+++ PooVersion2/GestionTextoPlano/src/main/java/Interfaces/ICursoArchivoDAO.java
@@ -1,3 +1,4 @@
+// ICursoArchivoDAO.java
 package Interfaces;
 
 import Modelos.Curso;
@@ -5,8 +6,8 @@ import java.util.List;
 
 public interface ICursoArchivoDAO {
     void addCurso(Curso curso);
-    Curso getCursoById(double id);
+    Curso getCursoById(Integer id);
     List<Curso> getAllCursos();
     void updateCurso(Curso curso);
-    void deleteCurso(double id);
+    void deleteCurso(Integer id);
 }
