+++ PooVersion2/GestionTextoPlano/src/main/java/Interfaces/IProfesorArchivoDAO.java
@@ -1,12 +1,12 @@
 package Interfaces;
 
-import Modelos.Persona;
+import Modelos.Profesor;
 import java.util.List;
 
 public interface IProfesorArchivoDAO {
     void addProfesor(Profesor profesor);
-    Profesor getProfesorById(int id);
+    Profesor getProfesorById(Integer id);
     List<Profesor> getAllProfesores();
     void updateProfesor(Profesor profesor);
-    void deleteProfesor(double id);
+    void deleteProfesor(Integer id);
 }
