+++ PooVersion2/GestionTextoPlano/src/main/java/Interfaces/IProgramaArchivoDAO.java
@@ -5,8 +5,8 @@ import java.util.List;
 
 public interface IProgramaArchivoDAO {
     void addPrograma(Programa programa);
-    Programa getProgramaById(double id);
+    Programa getProgramaById(Integer id);
     List<Programa> getAllProgramas();
     void updatePrograma(Programa programa);
-    void deletePrograma(double id);
+    void deletePrograma(Integer id);
 }
